feat: add 5 Pro AI tools, fix prices, i18n, and bug fixes
- Add 5 new Pro-exclusive Arena tools: Career Mentor, Company Prep,
  LinkedIn Generator, Cover Letter Generator, Salary Benchmark
  (full stack: controllers, services, DTOs, Angular components)
- Fix pricing: Starter €7.99/mo (€79.90/yr), Pro €15.99/mo (€159.90/yr)
- Change PricingTier price fields from int to double
- Enhanced pricing page: social proof bar, NEW badges, feature categories
- Contextual upsell on results page based on risk status (RED/YELLOW/GREEN)
- Fix language switcher: replace hardcoded English with translate pipe
- Fix navbar auth state: show Dashboard/Logout when authenticated
- Fix currency: all $ references changed to EUR
- Improve 404 page with branding and navigation
- Add validation hints to session wizards
- Fix backend 500s: null handling on career-risk-score, add GET marketplace/profile
- Add routes, command palette entries, tier feature flags for new tools

+++ studio-backend/src/main/java/ee/kerrete/ainterview/payment/dto/PricingTier.java
@@ -5,14 +5,14 @@ import java.util.List;
 public record PricingTier(
     String id,
     String name,
-    int price,
+    double price,
     String currency,
     List<String> features,
     boolean current,
     boolean popular,
     boolean subscription,
     String billingInterval,
-    Integer annualPrice,
+    Double annualPrice,
     Double annualMonthlyPrice,
     String badge
 ) {}
